Renamed and moved around node types
